Refactored Permissions Table into Component
